refactor: replace appNavigate with coreNavigate and streamline navigation in deeplink mutation
