+++ clients/desktop/src/vault/deposit/DepositForm/index.tsx
@@ -3,6 +3,7 @@ import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Opener } from '@lib/ui/base/Opener'
 import { Button } from '@lib/ui/buttons/Button'
+import { WithProgressIndicator } from '@lib/ui/flow/WithProgressIndicator'
 import { ChevronRightIcon } from '@lib/ui/icons/ChevronRightIcon'
 import { IconWrapper } from '@lib/ui/icons/IconWrapper'
 import { InputContainer } from '@lib/ui/inputs/InputContainer'
@@ -17,7 +18,6 @@ import { FieldValues, useForm } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 
 import { useAppPathParams } from '../../../navigation/hooks/useAppPathParams'
-import { WithProgressIndicator } from '../../keysign/shared/WithProgressIndicator'
 import { ChainAction } from '../ChainAction'
 import { useGetTotalAmountAvailableForChain } from '../hooks/useGetAmountTotalBalance'
 import { useGetMayaChainBondableAssetsQuery } from '../hooks/useGetMayaChainBondableAssetsQuery'
