+++ clients/desktop/src/vault/keysign/customMessage/SignCustomMessagePage.tsx
@@ -1,6 +1,7 @@
 import { create } from '@bufbuild/protobuf'
 import { CustomMessagePayloadSchema } from '@core/mpc/types/vultisig/keysign/v1/custom_message_payload_pb'
 import { FlowPageHeader } from '@lib/ui/flow/FlowPageHeader'
+import { WithProgressIndicator } from '@lib/ui/flow/WithProgressIndicator'
 import { TextInput } from '@lib/ui/inputs/TextInput'
 import { VStack } from '@lib/ui/layout/Stack'
 import { PageContent } from '@lib/ui/page/PageContent'
@@ -8,7 +9,6 @@ import { useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { StartKeysignPrompt } from '../components/StartKeysignPrompt'
-import { WithProgressIndicator } from '../shared/WithProgressIndicator'
 
 export const SignCustomMessagePage = () => {
   const { t } = useTranslation()
