+++ clients/desktop/src/vault/keysign/join/JoinKeysignPage.tsx
@@ -4,17 +4,17 @@ import { JoinMpcSessionFlow } from '@core/ui/mpc/session/join/JoinMpcSessionFlow
 import { CurrentHexEncryptionKeyProvider } from '@core/ui/mpc/state/currentHexEncryptionKey'
 import { IsInitiatingDeviceProvider } from '@core/ui/mpc/state/isInitiatingDevice'
 import { MpcSessionIdProvider } from '@core/ui/mpc/state/mpcSession'
+import { useCorePathState } from '@core/ui/navigation/hooks/useCorePathState'
 import { StepTransition } from '@lib/ui/base/StepTransition'
 import { useMemo } from 'react'
 
-import { useAppPathState } from '../../../navigation/hooks/useAppPathState'
 import { KeysignActionProvider } from '../action/KeysignActionProvider'
 import { KeysignServerUrlProvider } from './KeysignServerUrlProvider'
 import { KeysignVaultGuard } from './KeysignVaultGuard'
 import { JoinKeysignVerifyStep } from './verify/JoinKeysignVerifyStep'
 
 export const JoinKeysignPage = () => {
-  const { keysignMsg } = useAppPathState<'joinKeysign'>()
+  const { keysignMsg } = useCorePathState<'joinKeysign'>()
 
   const { sessionId, encryptionKeyHex } = keysignMsg
