+++ clients/desktop/src/vault/keysign/join/KeysignServerUrlProvider.tsx
@@ -2,6 +2,7 @@ import { FullPageFlowErrorState } from '@core/ui/flow/FullPageFlowErrorState'
 import { MpcServerTypeProvider } from '@core/ui/mpc/state/mpcServerType'
 import { MpcServerUrlProvider } from '@core/ui/mpc/state/mpcServerUrl'
 import { MpcPendingMessage } from '@core/ui/mpc/status/MpcPendingMessage'
+import { useCorePathState } from '@core/ui/navigation/hooks/useCorePathState'
 import { PageContent } from '@lib/ui/page/PageContent'
 import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderBackButton } from '@lib/ui/page/PageHeaderBackButton'
@@ -10,13 +11,12 @@ import { ChildrenProp } from '@lib/ui/props'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { useTranslation } from 'react-i18next'
 
-import { useAppPathState } from '../../../navigation/hooks/useAppPathState'
 import { useKeygenServerUrlQuery } from '../../keygen/server/queries/useKeygenServerUrlQuery'
 
 export const KeysignServerUrlProvider = ({ children }: ChildrenProp) => {
   const {
     keysignMsg: { serviceName, useVultisigRelay },
-  } = useAppPathState<'joinKeysign'>()
+  } = useCorePathState<'joinKeysign'>()
 
   const serverType = useVultisigRelay ? 'relay' : 'local'
