+++ clients/desktop/src/vault/keysign/join/KeysignVaultGuard.tsx
@@ -1,5 +1,6 @@
 import { FullPageFlowErrorState } from '@core/ui/flow/FullPageFlowErrorState'
 import { MpcLocalPartyIdProvider } from '@core/ui/mpc/state/mpcLocalPartyId'
+import { useCorePathState } from '@core/ui/navigation/hooks/useCorePathState'
 import { CurrentVaultProvider } from '@core/ui/vault/state/currentVault'
 import { useVaults } from '@core/ui/vault/state/vaults'
 import { getVaultId } from '@core/ui/vault/Vault'
@@ -9,12 +10,10 @@ import { useTranslation } from 'react-i18next'
 import { Link } from 'react-router-dom'
 
 import { makeAppPath } from '../../../navigation'
-import { useAppPathState } from '../../../navigation/hooks/useAppPathState'
 import { CurrentVaultCoinsProvider } from '../../state/currentVaultCoins'
 import { useCurrentVaultId } from '../../state/currentVaultId'
-
 export const KeysignVaultGuard = ({ children }: ChildrenProp) => {
-  const { vaultId } = useAppPathState<'joinKeysign'>()
+  const { vaultId } = useCorePathState<'joinKeysign'>()
   const vaults = useVaults()
 
   const { t } = useTranslation()
