+++ clients/desktop/src/vault/keysign/join/verify/JoinKeysignVerifyStep.tsx
@@ -1,5 +1,6 @@
 import { TxOverviewPanel } from '@core/ui/chain/tx/TxOverviewPanel'
 import { Button } from '@lib/ui/buttons/Button'
+import { WithProgressIndicator } from '@lib/ui/flow/WithProgressIndicator'
 import { PageContent } from '@lib/ui/page/PageContent'
 import { PageHeader } from '@lib/ui/page/PageHeader'
 import { PageHeaderBackButton } from '@lib/ui/page/PageHeaderBackButton'
@@ -7,8 +8,7 @@ import { PageHeaderTitle } from '@lib/ui/page/PageHeaderTitle'
 import { OnFinishProp } from '@lib/ui/props'
 import { useTranslation } from 'react-i18next'
 
-import { WithProgressIndicator } from '../../shared/WithProgressIndicator'
-import { KeysignTxOverview } from './KeysignTxOverview'
+import { JoinKeysignTxOverview } from './JoinKeysignTxOverview'
 
 export const JoinKeysignVerifyStep = ({ onFinish }: OnFinishProp) => {
   const { t } = useTranslation()
@@ -22,7 +22,7 @@ export const JoinKeysignVerifyStep = ({ onFinish }: OnFinishProp) => {
       <PageContent>
         <WithProgressIndicator value={0.6}>
           <TxOverviewPanel>
-            <KeysignTxOverview />
+            <JoinKeysignTxOverview />
           </TxOverviewPanel>
         </WithProgressIndicator>
         <Button onClick={onFinish}>{t('join_keysign')}</Button>
