+++ core/ui/navigation/index.ts
@@ -2,6 +2,7 @@ import { KeygenType } from '@core/mpc/keygen/KeygenType'
 import { KeysignMessagePayload } from '@core/mpc/keysign/keysignPayload/KeysignMessagePayload'
 import { KeygenMessage } from '@core/mpc/types/vultisig/keygen/v1/keygen_message_pb'
 import { ReshareMessage } from '@core/mpc/types/vultisig/keygen/v1/reshare_message_pb'
+import { KeysignMessage } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
 import { addQueryParams } from '@lib/utils/query/addQueryParams'
 import { withoutUndefinedFields } from '@lib/utils/record/withoutUndefinedFields'
 
@@ -18,6 +19,7 @@ export const corePaths = {
   reshareVault: '/vault/reshare',
   reshareVaultFast: '/vault/reshare/fast',
   reshareVaultSecure: '/vault/reshare/secure',
+  joinKeysign: '/join-keysign',
 } as const
 
 type CorePaths = typeof corePaths
@@ -36,6 +38,7 @@ export type CorePathState = {
     securityType: VaultSecurityType
     keysignPayload: KeysignMessagePayload
   }
+  joinKeysign: { vaultId: string; keysignMsg: KeysignMessage }
 }
 
 export type CorePathsWithParams = keyof CorePathParams
